+++ config/ajax-controller-options.php
@@ -1,0 +1,9 @@
+<?php
+return [
+    [
+        'key'         => 'controller_class',
+        'label'       => 'Controller Class Name',
+        'description' => 'Class name prefix for the AJAX controller, ie: Badges creates BadgesAjaxController',
+        'type'        => 'string',
+    ],
+];